# api.go
package main

import (
	"encoding/json"
	"log"
	"net/http"
)

type APIError struct {
	Code    int    `json:"code"`
	Message string `json:"message"`
}

func respondJSON[T any](w http.ResponseWriter, code int, result T) {
	// TODO: send empty slices instead of nil

	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(code)
	if err := json.NewEncoder(w).Encode(result); err != nil {
		log.Printf("Error writing json: %s", err)
	}
}

func respondError(w http.ResponseWriter, code int, msg string) {
	respondJSON(w, code, APIError{Code: code, Message: msg})
}

func internalServerError(w http.ResponseWriter, r *http.Request, err error) {
	respondError(w, http.StatusInternalServerError, "Internal server erro")
	logReqPrintf(r, "Internal server error: %s", err)
}