# docs/swagger.yaml
definitions:
  main.APIError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  main.AddPeerRequest:
    properties:
      allowedAddresses:
        description: |-
          List of CIDR subnets that are going to be assigned to the created peer.
          If null, then one address (/32 for v4, /128 v6) per each configured node's net
          will be assigned randomly.
        items:
          properties:
            mayOverlap:
              description: |-
                Subnet is not treated as exclusive for one peer, so it may overlap with other subnets,
                and it is not taken into account by random address allocation.
              type: boolean
            subnet:
              type: string
          type: object
        type: array
      endpoint:
        type: string
      owner:
        type: string
      persistent_keepalive:
        type: integer
      preshared_key_base64:
        type: string
      public_key_base64:
        description: omit to generate new random private key (not saved on the node)
        type: string
      random_preshared_key:
        description: |-
          If set, then a new random preshared key will be generated.
          This flag beign set is mutually exclusive with `preshared_key_base64` beign non-null.
        type: boolean
    type: object
  main.NewPeerResponse:
    properties:
      config:
        $ref: '#/definitions/main.WGQuickConfig'
      peer:
        $ref: '#/definitions/main.PeerResponse'
    type: object
  main.PeerResponse:
    properties:
      endpoint:
        type: string
      is_enabled:
        type: boolean
      owner:
        type: string
      persistent_keepalive:
        type: integer
      preshared_key_base64:
        type: string
      public_key_base64:
        type: string
    type: object
  main.ReserveSubnetRequest:
    properties:
      comment:
        description: Can be empty
        type: string
      may_overlap:
        type: boolean
      prefix:
        type: string
    type: object
  main.SubnetResponse:
    properties:
      comment:
        type: string
      id:
        type: integer
      may_overlap:
        type: boolean
      prefix:
        type: string
    type: object
  main.UpdatePeerRequest:
    properties:
      endpoint:
        type: string
      is_enabled:
        type: boolean
      owner:
        type: string
      persistent_keepalive:
        type: integer
      preshared_key_base64:
        type: string
    type: object
  main.WGQuickConfig:
    properties:
      interface:
        properties:
          addresses:
            description: CIDRs
            items:
              type: string
            type: array
          dns:
            description: based on configuration
            type: string
          mtu:
            description: based on configuration
            type: integer
          private_key:
            description: is set to a newly generated one only if no public key was
              provided in the request
            type: string
        type: object
      node_peer:
        properties:
          endpoint:
            type: string
          persistent_keepalive:
            description: based on configuration
            type: integer
          preshared_key:
            description: set only if preshared key was given, or random preshared
              key generation was requested
            type: string
          public_key:
            type: string
        type: object
    type: object
info:
  contact: {}
paths:
  /api/v1/peers:
    get:
      description: Returns a list of peers, optionally filtered by owner
      parameters:
      - description: Filter by owner
        in: query
        name: owner
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.PeerResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.APIError'
      summary: List peers
      tags:
      - peers
    post:
      consumes:
      - application/json
      description: Creates a new peer
      parameters:
      - description: peer
        in: body
        name: peer
        required: true
        schema:
          $ref: '#/definitions/main.AddPeerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.NewPeerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.APIError'
      summary: Add a new peer
      tags:
      - peers
  /api/v1/peers/{pubkey}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Peer key
        in: path
        name: key
        required: true
        type: string
      - description: Updated peer object
        in: body
        name: peer
        required: true
        schema:
          $ref: '#/definitions/main.UpdatePeerRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.APIError'
      summary: Update a peer
      tags:
      - peers
  /api/v1/subnets/{key}:
    delete:
      parameters:
      - description: Subnet id
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/main.APIError'
      summary: Delete a subnet
      tags:
      - subnets
  /api/v1/subnets/reserved:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.SubnetResponse'
            type: array
      summary: Get reserved subnets
      tags:
      - subnets
    post:
      consumes:
      - application/json
      parameters:
      - description: subnet
        in: body
        name: subnet
        required: true
        schema:
          $ref: '#/definitions/main.ReserveSubnetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.SubnetResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/main.APIError'
        "409":
          description: Subnet overlaps with existing
          schema:
            $ref: '#/definitions/main.APIError'
      summary: Reserve a subnet
      tags:
      - subnets
swagger: "2.0"